Move payment component into own payments component folder
