+++ react-app/src/pages/Payments.js
@@ -1,5 +1,5 @@
 // import components
-import Payment from '../components/Profile/Payment';
+import Payment from '../components/Payments';
 
 export default function Payments() {
   return (
